Issue #1441: Move LiveData tests to integration pkg.
The helpers package is for testing support code, not tests for support
code: all tests should go in integration/ or ui/.
