+++ app/src/androidTest/java/org/mozilla/tv/firefox/integration/LiveDataMapTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.helpers.ext
+package org.mozilla.tv.firefox.integration
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.Observer
